Fix for `-lGL` + `AUTO_NATIVE_LIBRARIES=0`
When calling `map_to_js_libs('GL')` we were mapping this to the JS
libraries, and not also including the native `libgl` library.  Normally
this doesn't matter since `libgl` is included by default, but not when
`AUTO_NATIVE_LIBRARIES=0` is used.

This also happens to fix an old bug (#10171) where we would look on the
file system for `libGL.a` before looking in `map_to_js_libs`.

Fixes: #10171
Fixes: #14335